Break out generic debugger functions to emu/utils
